big refactor from float to bigdecimal

+++ backend/moneyExchange/src/main/java/pl/niewadzj/moneyExchange/api/account/records/BalanceResponse.java
@@ -2,8 +2,10 @@ package pl.niewadzj.moneyExchange.api.account.records;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record BalanceResponse(Long currencyId,
                               String currencyCode,
-                              Float balance) {
+                              BigDecimal balance) {
 }
