+++ backend/moneyExchange/src/main/java/pl/niewadzj/moneyExchange/api/currency/records/CurrencyResponse.java
@@ -1,10 +1,11 @@
 package pl.niewadzj.moneyExchange.api.currency.records;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record CurrencyResponse(Long id,
                                String name,
                                String code,
                                LocalDateTime rateDate,
-                               double exchangeRate) {
+                               BigDecimal exchangeRate) {
 }
